requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
importlib-metadata==7.1.0
install==1.3.5
Markdown==3.6
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
zipp==3.18.1
